build: enforce zero compiler/linker warnings (#14)
Turns WSLDRIVE_WARNINGS_AS_ERRORS on by default and extends it to the linker
(/WX on MSVC, -Wl,--fatal-warnings on GCC). Fixes the remaining GCC
-Wsign-conversion warnings in the benchmarks via a scaled() helper and an
explicit reserve cast. First-party code (core/net/platform/agent/mount/tools/
tests/bench) now builds warning-clean under /W4 and -Wall -Wextra -Wpedantic
-Wconversion -Wsign-conversion -Wshadow -Wold-style-cast ...; third-party
FetchContent deps keep their own flags.

Verified: Windows /WX and Linux -Werror both build with zero warnings; 79/79
tests pass on both.

+++ bench/synthetic.hpp
@@ -22,7 +22,8 @@ struct SyntheticTree {
 
   static SyntheticTree build(int dirs, int sub, int files) {
     SyntheticTree s;
-    s.file_paths.reserve(static_cast<std::size_t>(dirs) * sub * files);
+    s.file_paths.reserve(static_cast<std::size_t>(dirs) * static_cast<std::size_t>(sub) *
+                         static_cast<std::size_t>(files));
     for (int d = 0; d < dirs; ++d) {
       const std::string dname = "module_" + std::to_string(d);
       const NodeId dn = *s.tree.insert(s.tree.root(), dname, kDirAttr);
